Move task and annotation to model
