+++ src/components/Viewer.vue
@@ -97,7 +97,7 @@ import Sidebar from '@/components/sidebar/Sidebar'
 import SelectSidebarItem from '@/components/sidebar/items/Select'
 import TranscribeSidebarItem from '@/components/sidebar/items/Transcribe'
 import BrowseSidebarItem from '@/components/sidebar/items/Browse'
-import Task from '@/task'
+import Task from '@/model/task'
 import drawOverlay from '@/utils/drawOverlay'
 import getImageUri from '@/utils/getImageUri'
 import extractRectFromImageUri from '@/utils/extractRectFromImageUri'
